feat(catalog): clean up after JanitorAI extraction and split the lorebook flow out of import
Extraction changes (JanitorWebViewProxy / JanitorExtractor):

- Bind a persona literally named `{{user}}` to the capture chat so JanitorAI's
  macro substitution is a no-op and the macro survives in the recovered card and
  lorebook entries. When the persona cannot be created, the account name that
  gets baked in is reported back and swapped to the macro instead.
- Restore `{{char}}` in every captured field (card, scenario, examples,
  greetings, lorebook text) by swapping the character's names back to the macro,
  longest name first. Names under 3 characters are left alone.
- Delete the chat, the persona and the throwaway proxy preset after every run,
  in dependency order, instead of leaving one dead chat per extraction behind.
- Read greetings from the chat's embedded character object, which carries them
  verbatim even when a closed card withholds them from /hampter/characters.
  Alternate greetings now survive the extraction path (they were dropped);
  the captured assistant turn is the last fallback for the primary greeting.
- Restore the account correctly in the two cases the old snapshot missed: an
  account with no proxy presets goes back to JLLM rather than staying in proxy
  mode on a deleted preset, and one with no readable generation settings gets a
  4096 context back instead of being left at the forced 0.
- Poll for the chat composer (up to 60s) instead of sleeping a fixed 2.5s before
  driving the input, and give the send button 30s to enable. Captures stay
  serialized — the run mutates shared account state.

Import flow:

- The import options sheet now offers character / lorebooks / character +
  lorebooks (`CatalogImportMode`).
- Lorebook capture moved out of the catalog preview into its own sheet, opened
  by the lorebook import modes and scoped to the imported character. It reuses
  the capture from the character import instead of running a second one.
- The preview's Lorebooks tab is now an inventory: what books the character has
  and of what kind, with a pointer to the Import button.

Tests: restoreMacros unit tests. Not run here — no Flutter SDK in this
environment.
