+++ assets/translations/en.json
@@ -1194,6 +1194,17 @@
   "catalog_import_char_only": "Character only",
   "catalog_import_char_only_hint": "Import the character without its lorebooks",
   "catalog_import_lorebooks_phase": "Importing lorebooks…",
+  "catalog_import_lorebooks_only": "Lorebooks only",
+  "catalog_import_lorebooks_only_hint": "Pull the lorebooks without adding the character",
+  "catalog_lorebooks_title": "Lorebooks",
+  "catalog_lorebooks_loading": "Loading lorebooks…",
+  "catalog_lorebooks_none": "No lorebooks attached.",
+  "catalog_lorebooks_public": "Public lorebooks",
+  "catalog_lorebooks_closed": "Closed lorebooks",
+  "catalog_lorebook_kind_public": "{} entries · downloads whole",
+  "catalog_lorebook_kind_scripted": "Scripted (advanced) · rebuilt with your LLM",
+  "catalog_lorebook_kind_closed": "Private · recovered from the prompt",
+  "catalog_lorebooks_import_hint": "Lorebooks are pulled from the Import button: pick \"Lorebooks\" or \"Character + lorebooks\".",
   "catalog_extract_via_dc": "Extract via DataCat",
   "catalog_imported": "Imported!",
   "catalog_imported_desc": "{name} has been added to your library.",
