+++ assets/translations/ru.json
@@ -1259,6 +1259,17 @@
   "catalog_import_char_only": "Только персонаж",
   "catalog_import_char_only_hint": "Импортировать персонажа без лорбуков",
   "catalog_import_lorebooks_phase": "Импорт лорбуков…",
+  "catalog_import_lorebooks_only": "Только лорбуки",
+  "catalog_import_lorebooks_only_hint": "Забрать лорбуки, не добавляя персонажа",
+  "catalog_lorebooks_title": "Лорбуки",
+  "catalog_lorebooks_loading": "Загрузка лорбуков…",
+  "catalog_lorebooks_none": "Лорбуков нет.",
+  "catalog_lorebooks_public": "Публичные лорбуки",
+  "catalog_lorebooks_closed": "Закрытые лорбуки",
+  "catalog_lorebook_kind_public": "Записей: {} · скачивается целиком",
+  "catalog_lorebook_kind_scripted": "Скриптовый (advanced) · пересобирается вашей LLM",
+  "catalog_lorebook_kind_closed": "Приватный · восстанавливается из промпта",
+  "catalog_lorebooks_import_hint": "Лорбуки забираются через кнопку импорта: выберите «Только лорбуки» или «Персонаж + лорбуки».",
   "catalog_extract_via_dc": "Извлечь через DataCat",
   "catalog_imported": "Импортировано!",
   "catalog_imported_desc": "{name} добавлен в библиотеку.",
